# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 'Autpost API'
  description: 'API documentation for the Autpost application'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:9527/api'
    description: 'Local development server'
paths:
...
  /user/avatar:
    post:
      tags:
        - User
      summary: 'Upload user avatar'
      operationId: 93c05b6d6ba56d027806ca7563e9014b
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
              type: object
      responses:
        200:
          description: 'Avatar uploaded successfully'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - User
      summary: 'Delete user avatar'
      operationId: c0cce51e99fca4cf8d48edf2bcbc9a48
      responses:
        200:
          description: 'Avatar deleted successfully'
        401:
          description: Unauthenticated
      security:
        -
          bearerAuth: []
  /user/profile:
    patch:
      tags:
        - User
      summary: 'Update user profile'
      operationId: b6f4aedd5f5cc1f82d0a13f2395a12b5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 255
              type: object
      responses:
        200:
          description: 'Profile updated successfully'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
  /user/email/request:
    post:
      tags:
        - User
      summary: 'Request email change - sends verification code to new email'
      operationId: e1c63330246b0305d530f459fb99b5d5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
              type: object
      responses:
        200:
          description: 'Verification code sent'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
  /user/email/verify:
    post:
      tags:
        - User
      summary: 'Verify email change code and update email'
      operationId: 9a0aa42cd01ab76af20d6a6f1f0701b8
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - code
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  maxLength: 6
                  minLength: 6
              type: object
      responses:
        200:
          description: 'Email updated successfully'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
  /user/password:
    post:
      tags:
        - User
      summary: 'Change user password'
      operationId: 2219563b052b225b4f25b74e12552cac
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - current_password
                - password
                - password_confirmation
              properties:
                current_password:
                  type: string
                password:
                  type: string
                  minLength: 8
                password_confirmation:
                  type: string
              type: object
      responses:
        200:
          description: 'Password changed successfully'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Enter your Sanctum token'
      bearerFormat: token
      scheme: bearer
...